requirements.txt
SQLAlchemy~=2.0.36
uvicorn~=0.32.1
fastapi~=0.115.5
PyJWT~=2.10.0
pydantic~=2.10.0
passlib~=1.7.4